Module sykl::api

Constants

CLIENT_IDENTIFIER
The Client-Identifier of this application.

Functions

get
Sends a get request to a specific url with the Client-Identifier header set to the application id.

information
Sends a get request to https://gbfs.urbansharing.com/oslobysykkel.no/station_information.json, parses the response as json and deserializes the json response into a bikeshare struct using serde.

status
Sends a get request to https://gbfs.urbansharing.com/oslobysykkel.no/station_status.json, parses the response as json and deserializes the json response into a bikeshare struct using serde.