Module sykl::formatting

Structs

Row
This struct represents a row in the pretty printed output.

Constants

EXTRA_PADDING
The extra padding to add in addition to the length of the longest value of the column.

HEADER
The headers used in the pretty print.

PLACEHOLDER
The string to use in the case where no status information is available for a station.

Functions

pretty_print_stations
Prints each stations name, number of available docks and number of available bikes in 3 columns per row with a header row.